エンティティの作成を個々のファクトリでやってたけど、共通処理ばかりだし、jsonデータから生成してるからまったく無駄なことをしていたので、createEntitiesに統一,systemをインターフェイスで管理できるようにした。

+++ Game/source/engine/Components/sprite_anim.hpp
@@ -9,7 +9,7 @@ namespace sdl_engine
    struct SpriteAnim
    {
       entt::resource<SpriteAnimResource> sprite_anim;
-      f32                                current_frame {};
+      u32                                current_frame {};
       f64                                elapsed_time {};    // •b’PˆÊ
       bool                               is_playing {};
    };
